+++ packages/cli/src/util/target/types.ts
@@ -1,0 +1,10 @@
+export interface CustomEnvironment {
+  id: string;
+  slug: string;
+  createdAt: number;
+  updatedAt: number;
+  type: string;
+  description: string;
+  name: string;
+  domains: [];
+}